requirements.txt
neo4j==5.14.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
fastapi==0.104.1
uvicorn==0.24.0
openai==1.3.5
httpx==0.24.1

# GraphRAG dependencies
numpy>=1.21.0
sentence-transformers>=2.2.0

# Testing dependencies
pytest-cov==4.1.0